:truck: Move creation of context resources into 'shm deploy'
